server/cutecut/requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
docutils==0.19
freelancersdk==0.1.20
guzzle-sphinx-theme==0.7.11
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
snowballstemmer==2.2.0
Sphinx==6.0.0b2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
urllib3==1.26.13
